+++ includes/admin/admin-branding.php
@@ -43,7 +43,7 @@ function replace_login_logo() {
 
 	?><style type="text/css">
 		body.login h1 a {
-			background-image: url(<?php echo get_stylesheet_directory_uri(); ?>/build/images/login-logo.svg);
+			background-image: url(<?php echo get_stylesheet_directory_uri(); ?>/images/login-logo.svg);
 
 			/* Adjust to the dimensions of your logo. WP Default: 84px 84px */
 			background-size: 84px 84px;